committ #15
cleaned up the signUp page as well as made the search page look better and work similarly to instagram. just need to add the userProfile component for that to allow the user to follow a link trail to another users profile to add them as a friend or invite them to a group!