commit #23
design changes for home and group pages. began the journey on pfp's can cache them but still cannot storage them via the firebase storage tool until i put in my credit card, unfortunately. It's just time. there are of course other things I could work on but this is the most pressing matter at this stage of the semester. It's time to add photos and videos.